+++ gax-java/gax-httpjson/src/main/java/com/google/api/gax/httpjson/ManagedHttpJsonInterceptorChannel.java
@@ -29,11 +29,9 @@
  */
 package com.google.api.gax.httpjson;
 
-import com.google.api.core.BetaApi;
 import com.google.common.annotations.VisibleForTesting;
 import java.util.concurrent.TimeUnit;
 
-@BetaApi
 class ManagedHttpJsonInterceptorChannel extends ManagedHttpJsonChannel {
 
   private final ManagedHttpJsonChannel channel;