+++ gax-java/gax-httpjson/src/main/java/com/google/api/gax/httpjson/longrunning/OperationsSettings.java
@@ -32,7 +32,6 @@ package com.google.api.gax.httpjson.longrunning;
 import static com.google.api.gax.httpjson.longrunning.OperationsClient.ListOperationsPagedResponse;
 
 import com.google.api.core.ApiFunction;
-import com.google.api.core.BetaApi;
 import com.google.api.gax.core.GoogleCredentialsProvider;
 import com.google.api.gax.core.InstantiatingExecutorProvider;
 import com.google.api.gax.httpjson.InstantiatingHttpJsonChannelProvider;
@@ -84,7 +83,6 @@ import java.util.List;
  * OperationsSettings operationsSettings = operationsSettingsBuilder.build();
  * }</pre>
  */
-@BetaApi
 public class OperationsSettings extends ClientSettings<OperationsSettings> {
 
   /** Returns the object with the settings used for calls to listOperations. */